Create we do what others do
One puzzle was put in the friends WhatsApp group . All were trying to solve it. I just tried for some time. I had other works too. One of my close friends without much mathematical knowledge was posting so many answers. Then I found all were forwarded messages. So I sensed this man tried forwading the puzzle to various other groups to get some answer . He got some answers from people and he was very confidently posting it. I found out that and was teasing him. Before I found that, even I forwarded the puzzle to some 5 groups fo find the answer. If some one confidently posting some answer, I thought I will check and put, so that our close friends in the group appreciate me. There were many answers in other groups. I was still finding for a confident and majority one of it. I didn't get still. Meanwhile lot of people from other groups were asking me whether what they posted was right. They were asking the real answer too. They thought I only manufactured the puzzle and were asking the genuine answer from me . Then in our main friend group where the original puzzle was put , lot of answers came and one answer was confidently posted. Lot of people were acknowledging it as the correct answer. So almost every one thought it was the correct answer. Then we asked the friend who made the puzzle whether that was right. He was quiet for sometime. Then we gave so much pressure to make him finalize the answer. then he said sarcastically this( yoah.. iruyya.. naane idha 10 gropukku forward panniyirukken...) keep quiet for some time. Even I have forwarded the puzzle to lot of groups. your finalized answer was not not a majority no answer has come as a majority... let all send... I will find out the majority and say... we were shockingly laughed and the laugh was unstoppable. 

We learnt from this that this entire life is based on forwards and what others say. No one is trying to find the reality and the real ones. All are living mostly on what others say and do ... it was only the majority of the life style and even if it's wrong thought due to its majority status, we are following it. Mostly we follow what others live and based on what others say, even if it's wrong. If majority of them think right, we are lucky... if majority of them think wrong, we follow them not to become right... this is what even our political system of most of the countries in the world is following. If people with wrong thoughts & wrong deeds become a majority , then wrong becomes the right to rule and make people's life wrong